Fix new themes not playing after refresh
When themes were refreshed, device.themes was replaced with a new
IndexList, but session_manager still held a reference to the old one.
New themes wouldn't be found when trying to play.

Added set_themes() method to session_manager and call it after
theme refresh to update the reference.

Version 1.1.11-dev